Allow using penalty function instead of object function
The penalty function includes a special special term for infeasible
solutions. For feasible solutions, the penalty function is the objective
function - for infeasible solutions it adds a sum based on constraint
functions to the objective function of the worst feasible solution. To
compute the latter, the penalty function receives an additional
parameter: the current population.