Parameterised functions and removed getIndividual
The `getIndividual` function was performing unnecessary work, so is
factored out in this commit. Instead, a sample individual is passed
as a parameter. This has the added benefit of making it clear to a
reader what kind of data is being worked with.

Note: n-queens P increased to 200 due to flaky test.

of accessing members