Increased piping performance
The previous prototype was quite slow due to using remote channels and the pipe reading not being made
in the process it was running. Now the remote channels are gone and the pipe reading and the objective
function are determined in the same process, which is much more efficient and faster.