Started implementing boundary functionalities
Before this branch the variable boundaries were checked inside the objective function, returning an Inf value when they were not inside boundaries. This was not ideal, since some individuals did not have a chance to improve in further iterations because they were automatically discarded.
Now the boundary checking is made inside the mutation functions. In this way, all the individuals have genes that respect the boundaries, reducing the number of objective function evaluations that would be Inf.